LogViewer: stay scrolled to bottom
Follow the log if we're already scrolled to the bottom. But if the user
isn't scrolled all the way, stop tailing.

While we're at it, quit touching arrays from two threads at the same
time by doing the log addition from inside the main context. This also
allows better synchronization for seeing whether or not we're
scrolled.